[Feature] Update balance display
- Creates FiaValueDetails component
- Updates balance display and header  on CurrencyInputPanel
