+++ src/components/CurrencyInputPanel/index.tsx
@@ -1,4 +1,4 @@
-import { Currency, CurrencyAmount, Pair, Percent, JSBI } from '@swapr/sdk'
+import { Currency, CurrencyAmount, Pair, Percent } from '@swapr/sdk'
 import React, { useState, useCallback } from 'react'
 import styled from 'styled-components'
 import { useCurrencyBalance } from '../../state/wallet/hooks'
@@ -12,7 +12,7 @@ import { ReactComponent as DropDown } from '../../assets/images/dropdown.svg'
 
 import { useActiveWeb3React } from '../../hooks'
 import { useTranslation } from 'react-i18next'
-import { warningFiatSeverity } from '../../utils/prices'
+import FiatValueDetails from '../FiatValueDetails'
 
 const InputRow = styled.div<{ selected: boolean }>`
   ${({ theme }) => theme.flexRowNoWrap}
@@ -75,7 +75,7 @@ const Container = styled.div<{ focused: boolean }>`
   border: solid 1px ${({ focused, theme }) => (focused ? theme.bg3 : theme.bg1And2)};
   border-radius: 12px;
   transition: border 0.3s ease;
-  padding: 17px 22px;
+  padding: 18px 22px;
 `
 
 const Content = styled.div``
@@ -106,11 +106,6 @@ const UppercaseHelper = styled.span`
 
 const FiatRow = styled.div``
 
-const PriceImpactText = styled.span<{ warning?: boolean }>`
-  color: ${({ theme, warning }) => warning && theme.red1};
-  margin-left: 8px;
-`
-
 interface CurrencyInputPanelProps {
   value: string
   onUserInput: (value: string) => void
@@ -138,7 +133,7 @@ export default function CurrencyInputPanel({
   onUserInput,
   onMax,
   showMaxButton,
-  label = 'Input',
+  label,
   onCurrencySelect,
   currency,
   disableCurrencySelect = false,
@@ -174,40 +169,16 @@ export default function CurrencyInputPanel({
     setFocused(false)
   }, [])
 
-  const fiatPriceImpactSeverity = warningFiatSeverity(priceImpact)
-
   return (
     <InputPanel id={id}>
       <Container focused={focused}>
         <Content>
-          {!hideInput && (
+          {!hideInput && label && (
             <LabelRow>
               <RowBetween>
                 <TYPE.body fontWeight="600" fontSize="11px" lineHeight="13px" letterSpacing="0.08em">
                   <UppercaseHelper>{label}</UppercaseHelper>
                 </TYPE.body>
-                {account && (
-                  <TYPE.body
-                    onClick={onMax}
-                    fontWeight="600"
-                    fontSize="11px"
-                    lineHeight="13px"
-                    letterSpacing="0.08em"
-                    style={{
-                      display: 'inline',
-                      cursor:
-                        !hideBalance && !!(currency || pair) && (balance || selectedCurrencyBalance)
-                          ? 'pointer'
-                          : 'auto'
-                    }}
-                  >
-                    <UppercaseHelper>
-                      {!hideBalance && !!(currency || pair) && (balance || selectedCurrencyBalance)
-                        ? (customBalanceText ?? 'Balance: ') + (balance || selectedCurrencyBalance)?.toSignificant(6)
-                        : '-'}
-                    </UppercaseHelper>
-                  </TYPE.body>
-                )}
               </RowBetween>
             </LabelRow>
           )}
@@ -264,14 +235,32 @@ export default function CurrencyInputPanel({
           </InputRow>
           <FiatRow>
             <RowBetween>
-              <TYPE.body fontWeight="600" fontSize="11px" lineHeight="13px" letterSpacing="0.08em">
-                <UppercaseHelper>${fiatValue ? fiatValue.toFixed(2) : '0'}</UppercaseHelper>
-                {priceImpact && (
-                  <PriceImpactText warning={fiatPriceImpactSeverity === 1}>
-                    {priceImpact.multiply(JSBI.BigInt(-100)).toSignificant(3)}%
-                  </PriceImpactText>
-                )}
-              </TYPE.body>
+              <FiatValueDetails fiatValue={fiatValue} priceImpact={priceImpact} />
+              {account && (
+                <TYPE.body
+                  onClick={onMax}
+                  fontWeight="600"
+                  fontSize="10px"
+                  lineHeight="12px"
+                  letterSpacing="0.08em"
+                  style={{
+                    display: 'inline',
+                    cursor:
+                      !hideBalance && !!(currency || pair) && (balance || selectedCurrencyBalance) ? 'pointer' : 'auto'
+                  }}
+                >
+                  <UppercaseHelper>
+                    {!hideBalance && !!(currency || pair) && (balance || selectedCurrencyBalance) && (
+                      <>
+                        {customBalanceText ?? 'Balance: '}
+                        <TYPE.small as="span" fontWeight="600" color="text3" style={{ textDecoration: 'underline' }}>
+                          {(balance || selectedCurrencyBalance)?.toSignificant(6)}
+                        </TYPE.small>
+                      </>
+                    )}
+                  </UppercaseHelper>
+                </TYPE.body>
+              )}
             </RowBetween>
           </FiatRow>
         </Content>
